refactor(staking)
