+++ frontend/src/app/(routes)/staking/components/DialogAllValidators.tsx
@@ -1,57 +1,25 @@
-import { StakingMenuAction, Validator, Validators } from '@/types/staking';
-import {
-  canDelegate,
-  getValidatorRank,
-  getValidatorStatus,
-} from '@/utils/util';
-import { Dialog, DialogContent, Pagination, Tooltip } from '@mui/material';
+import { StakingMenuAction, Validators } from '@/types/staking';
+import { Dialog, DialogContent } from '@mui/material';
 import Image from 'next/image';
-import React, { useEffect, useState } from 'react';
-import ValidatorLogo from './ValidatorLogo';
+import React, { useState } from 'react';
+import ActiveValidators from './ActiveValidators';
+import InactiveValidators from './InactiveValidators';
 
 type HandleClose = () => void;
 
-const paginationComponentStyles = {
-  '& .MuiPaginationItem-page': {
-    color: '#fff',
-    '&:hover': {
-      backgroundColor: '#ffffff1a',
-    },
-    fontSize: '12px',
-    minWidth: '24px',
-    height: '24px',
-    borderRadius: '4px',
-  },
-  '& .Mui-selected': {
-    background: 'linear-gradient(180deg, #4AA29C 0%, #8B3DA7 100%)',
-    '&:hover': {
-      opacity: '0.95',
-    },
-  },
-  '& .MuiPaginationItem-icon': {
-    color: '#fff',
-  },
-  '&.Mui-disabled': {
-    color: 'red',
-  },
-  '& .MuiPaginationItem-ellipsis, & .MuiPaginationItem-ellipsisIcon': {
-    color: 'white',
-  },
-};
-
-const PER_PAGE = 7;
+interface DialogAllValidatorsProps {
+  handleClose: HandleClose;
+  open: boolean;
+  validators: Validators;
+  onMenuAction: StakingMenuAction;
+}
 
 const DialogAllValidators = ({
   handleClose,
   open,
   validators,
   onMenuAction,
-}: {
-  handleClose: HandleClose;
-  open: boolean;
-  validators: Validators;
-  onMenuAction: StakingMenuAction;
-}) => {
+}: DialogAllValidatorsProps) => {
   const [searchTerm, setSearchTerm] = useState('');
   const [active, setActive] = useState<boolean>(true);
 
@@ -156,368 +124,3 @@ const DialogAllValidators = ({
 };
 
 export default DialogAllValidators;
-
-const ActiveValidators = ({
-  validators,
-  searchTerm,
-  onMenuAction,
-}: {
-  validators: Validators;
-  searchTerm: string;
-  onMenuAction: StakingMenuAction;
-}) => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const [slicedValidators, setSlicedValidators] = useState<string[]>([]);
-  const [filtered, setFiltered] = useState<string[]>([]);
-
-  useEffect(() => {
-    if (validators?.activeSorted.length < PER_PAGE) {
-      setSlicedValidators(validators?.activeSorted);
-    } else {
-      setCurrentPage(1);
-      setSlicedValidators(validators?.activeSorted?.slice(0, 1 * PER_PAGE));
-    }
-  }, [validators?.activeSorted]);
-
-  useEffect(() => {
-    const filteredValidators = validators?.activeSorted.filter(
-      (validator) =>
-        validators.active[validator]?.description.moniker
-          .toLowerCase()
-          .includes(searchTerm.toLowerCase())
-    );
-    setFiltered(filteredValidators);
-    setCurrentPage(1);
-  }, [searchTerm, validators?.activeSorted]);
-  return (
-    <>
-      <div className="px-10">
-        {searchTerm.length ? (
-          <>
-            <Filtered
-              filtered={filtered}
-              validators={validators}
-              active={true}
-              onMenuAction={onMenuAction}
-            />
-          </>
-        ) : (
-          <>
-            <div className="flex flex-col gap-6">
-              {slicedValidators?.map((validator, index) => {
-                const moniker =
-                  validators.active[validator]?.description.moniker;
-                const identity =
-                  validators.active[validator]?.description.identity;
-                const commission =
-                  Number(
-                    validators.active[validator]?.commission?.commission_rates
-                      .rate
-                  ) * 100;
-                const jailed = validators.active[validator]?.jailed;
-                const status = validators.active[validator]?.status;
-                const rank = getValidatorRank(
-                  validator,
-                  validators.activeSorted
-                );
-
-                return (
-                  <ValidatorComponent
-                    key={index + PER_PAGE * (currentPage - 1)}
-                    moniker={moniker}
-                    identity={identity}
-                    commission={commission}
-                    jailed={jailed}
-                    status={status}
-                    active={true}
-                    rank={rank}
-                    onMenuAction={onMenuAction}
-                    validator={validators.active[validator]}
-                  />
-                );
-              })}
-            </div>
-            <div className="w-full h-[0.25px] bg-[#FFFFFF66] my-6"></div>
-            <div className="absolute bottom-12 right-10">
-              <Pagination
-                sx={paginationComponentStyles}
-                count={Math.ceil(validators?.activeSorted?.length / PER_PAGE)}
-                shape="circular"
-                onChange={(_, v) => {
-                  setCurrentPage(v);
-                  setSlicedValidators(
-                    validators?.activeSorted?.slice(
-                      (v - 1) * PER_PAGE,
-                      v * PER_PAGE
-                    )
-                  );
-                }}
-              />
-            </div>
-          </>
-        )}
-      </div>
-    </>
-  );
-};
-
-const InactiveValidators = ({
-  validators,
-  searchTerm,
-  onMenuAction,
-}: {
-  validators: Validators;
-  searchTerm: string;
-  onMenuAction: StakingMenuAction;
-}) => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const [slicedValidators, setSlicedValidators] = useState<string[]>([]);
-  const [filtered, setFiltered] = useState<string[]>([]);
-
-  useEffect(() => {
-    if (validators?.activeSorted.length < PER_PAGE) {
-      setSlicedValidators(validators?.inactiveSorted);
-    } else {
-      setCurrentPage(1);
-      setSlicedValidators(validators?.inactiveSorted?.slice(0, 1 * PER_PAGE));
-    }
-  }, [validators?.inactiveSorted]);
-
-  useEffect(() => {
-    const filteredValidators = validators?.inactiveSorted.filter(
-      (validator) =>
-        validators.inactive[validator]?.description.moniker
-          .toLowerCase()
-          .includes(searchTerm.toLowerCase())
-    );
-    setFiltered(filteredValidators);
-    setCurrentPage(1);
-  }, [searchTerm, validators?.inactiveSorted]);
-  return (
-    <>
-      <div className="px-10">
-        {searchTerm.length ? (
-          <>
-            <Filtered
-              filtered={filtered}
-              validators={validators}
-              active={false}
-              onMenuAction={onMenuAction}
-            />
-          </>
-        ) : (
-          <>
-            <div className="flex flex-col gap-6">
-              {slicedValidators?.map((validator, index) => {
-                const moniker =
-                  validators.inactive[validator]?.description.moniker;
-                const identity =
-                  validators.inactive[validator]?.description.identity;
-                const commission =
-                  Number(
-                    validators.inactive[validator]?.commission?.commission_rates
-                      .rate
-                  ) * 100;
-                const jailed = validators.inactive[validator]?.jailed;
-                const status = validators.inactive[validator]?.status;
-                const rank = getValidatorRank(validator, [
-                  ...validators.activeSorted,
-                  ...validators.inactiveSorted,
-                ]);
-
-                return (
-                  <ValidatorComponent
-                    key={index + PER_PAGE * (currentPage - 1)}
-                    moniker={moniker}
-                    identity={identity}
-                    commission={commission}
-                    jailed={jailed}
-                    status={status}
-                    active={false}
-                    rank={rank}
-                    onMenuAction={onMenuAction}
-                    validator={validators.inactive[validator]}
-                  />
-                );
-              })}
-            </div>
-            <div className="w-full h-[0.25px] bg-[#FFFFFF66] my-6"></div>
-            <div className="absolute bottom-12 right-10">
-              <Pagination
-                sx={paginationComponentStyles}
-                count={Math.ceil(validators?.inactiveSorted?.length / PER_PAGE)}
-                shape="circular"
-                onChange={(_, v) => {
-                  setCurrentPage(v);
-                  setSlicedValidators(
-                    validators?.inactiveSorted?.slice(
-                      (v - 1) * PER_PAGE,
-                      v * PER_PAGE
-                    )
-                  );
-                }}
-              />
-            </div>
-          </>
-        )}
-      </div>
-    </>
-  );
-};
-
-const ValidatorComponent = ({
-  moniker,
-  identity,
-  commission,
-  jailed,
-  status,
-  active,
-  rank,
-  onMenuAction,
-  validator,
-}: {
-  moniker: string;
-  identity: string;
-  commission: number;
-  jailed: boolean;
-  status: string;
-  active: boolean;
-  rank: string;
-  onMenuAction: StakingMenuAction;
-  validator: Validator;
-}) => {
-  const validatorStatus = getValidatorStatus(jailed, status);
-  const enableDelegate = canDelegate(validatorStatus);
-  return (
-    <div className="flex justify-between items-center txt-sm text-white font-normal">
-      <div className="flex gap-4 items-center">
-        <div className="bg-[#fff] rounded-full">
-          <ValidatorLogo identity={identity} width={40} height={40} />
-        </div>
-        <div className="flex flex-col gap-2 w-[200px]">
-          <div className="flex gap-2 items-center cursor-default">
-            <Tooltip title={moniker} placement="top">
-              <div className="text-[14px] leading-normal truncate">
-                {moniker}
-              </div>
-            </Tooltip>
-          </div>
-        </div>
-      </div>
-      <div className="leading-3">{rank}</div>
-      <div className="leading-3 min-w-[132px] text-center">
-        {commission ? String(commission.toFixed(0)) + '%' : '-'} Commission
-      </div>
-      {active ? null : (
-        <div className="min-w-[102px] text-center leading-3">
-          {validatorStatus}
-        </div>
-      )}
-      <div>
-        <button
-          className={
-            enableDelegate
-              ? `delegate-button`
-              : `delegate-button delegate-button-inactive`
-          }
-          onClick={() => {
-            if (enableDelegate) {
-              onMenuAction('delegate', validator);
-            }
-          }}
-        >
-          Delegate
-        </button>
-      </div>
-    </div>
-  );
-};
-
-const Filtered = ({
-  filtered,
-  validators,
-  active,
-  onMenuAction,
-}: {
-  filtered: string[];
-  validators: Validators;
-  active: boolean;
-  onMenuAction: StakingMenuAction;
-}) => {
-  const [slicedValidators, setSlicedValidators] = useState<string[]>([]);
-
-  const [currentPage, setCurrentPage] = useState(1);
-
-  useEffect(() => {
-    if (filtered.length < PER_PAGE) {
-      setSlicedValidators(filtered);
-    } else {
-      setCurrentPage(1);
-      setSlicedValidators(filtered?.slice(0, 1 * PER_PAGE));
-    }
-  }, [filtered]);
-
-  return (
-    <>
-      {slicedValidators.length ? (
-        <>
-          <div className="flex flex-col gap-6">
-            {slicedValidators?.map((validator, index) => {
-              let validatorsSet;
-              let rank;
-              if (active) {
-                validatorsSet = validators.active;
-                rank = getValidatorRank(validator, validators.activeSorted);
-              } else {
-                validatorsSet = validators.inactive;
-                rank = getValidatorRank(validator, [
-                  ...validators.activeSorted,
-                  ...validators.inactiveSorted,
-                ]);
-              }
-              const moniker = validatorsSet[validator]?.description.moniker;
-              const identity = validatorsSet[validator]?.description.identity;
-              const commission =
-                Number(
-                  validatorsSet[validator]?.commission?.commission_rates.rate
-                ) * 100;
-              const jailed = validatorsSet[validator]?.jailed;
-              const status = validatorsSet[validator]?.status;
-
-              return (
-                <ValidatorComponent
-                  key={index + PER_PAGE * (currentPage - 1)}
-                  moniker={moniker}
-                  identity={identity}
-                  commission={commission}
-                  jailed={jailed}
-                  status={status}
-                  active={active}
-                  rank={rank}
-                  onMenuAction={onMenuAction}
-                  validator={validatorsSet[validator]}
-                />
-              );
-            })}
-          </div>
-          <div className="w-full h-[0.25px] bg-[#FFFFFF66] my-6"></div>
-          <div className="absolute bottom-12 right-10">
-            <Pagination
-              sx={paginationComponentStyles}
-              count={Math.ceil(filtered?.length / PER_PAGE)}
-              shape="circular"
-              onChange={(_, v) => {
-                setCurrentPage(v);
-                setSlicedValidators(
-                  filtered?.slice((v - 1) * PER_PAGE, v * PER_PAGE)
-                );
-              }}
-            />
-          </div>
-        </>
-      ) : (
-        <div className="text-white text-center">No validators found</div>
-      )}
-    </>
-  );
-};
