+++ frontend/src/app/(routes)/staking/styles.ts
@@ -1,0 +1,27 @@
+export const paginationComponentStyles = {
+  '& .MuiPaginationItem-page': {
+    color: '#fff',
+    '&:hover': {
+      backgroundColor: '#ffffff1a',
+    },
+    fontSize: '12px',
+    minWidth: '24px',
+    height: '24px',
+    borderRadius: '4px',
+  },
+  '& .Mui-selected': {
+    background: 'linear-gradient(180deg, #4AA29C 0%, #8B3DA7 100%)',
+    '&:hover': {
+      opacity: '0.95',
+    },
+  },
+  '& .MuiPaginationItem-icon': {
+    color: '#fff',
+  },
+  '&.Mui-disabled': {
+    color: 'red',
+  },
+  '& .MuiPaginationItem-ellipsis, & .MuiPaginationItem-ellipsisIcon': {
+    color: 'white',
+  },
+};
